requirements.txt
# Deep Learning Framework
torch>=1.11.0
torchvision>=0.12.0
tensorboard>=2.9.1

# Computer Vision and Image Processing
opencv-python>=4.7.0
Pillow>=9.1.1
ImageHash>=4.3.1
PyWavelets>=1.4.1

# Data Processing and Analysis
numpy>=1.22.4
pandas>=2.0.2
scikit-learn>=1.3.1
scipy>=1.10.1

# Visualization
matplotlib>=3.5.2
seaborn>=0.12.2

# YOLOv5 Related
ultralytics>=8.0.0
thop>=0.1.1

# Feature Extraction and Model Interpretation
gradio>=4.38.1
transformers>=4.27.1

# File and Data Processing
openpyxl>=3.1.5
PyYAML>=6.0
filelock>=3.12.4
fsspec>=2023.9.2

# Progress and Parallel Processing
tqdm>=4.65.0
joblib>=1.3.2

# Jupyter Notebook and Interactive Components
ipykernel>=6.15.0
jupyter>=1.0.0
ipywidgets>=7.7.0
nbformat>=5.4.0

# Utility Tools
requests>=2.32.3
regex>=2023.8.8